+++ src/main/java/szewek/mcflux/wrapper/MCFluxWrapper.java
@@ -11,15 +11,16 @@ import szewek.mcflux.util.MCFluxLocation;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 public final class MCFluxWrapper implements ICapabilityProvider, INBTSerializable<NBTBase> {
 	public static final MCFluxLocation MCFLUX_WRAPPER = new MCFluxLocation("wrapper");
 	final Object mainObject;
-	private ICapabilityProvider[] providers = null;
-	private INBTSerializable<NBTBase>[] serializables = null;
-	private String[] names = null;
+	private ICapabilityProvider[] providers = new ICapabilityProvider[0];
+	private Map<String, INBTSerializable<NBTBase>> nbts = new HashMap<>();
+	private NBTBase cachedNBT = null;
 
 	MCFluxWrapper(Object o) {
 		mainObject = o;
@@ -35,8 +36,9 @@ public final class MCFluxWrapper implements ICapabilityProvider, INBTSerializabl
 		return false;
 	}
 
+	@SuppressWarnings("unchecked")
 	@Nullable @Override public <T> T getCapability(@Nonnull Capability<T> cap, @Nullable EnumFacing f) {
-		if (providers == null)
+		if (providers.length == 0)
 			return null;
 		T t;
 		for (ICapabilityProvider icp : providers) {
@@ -48,53 +50,48 @@ public final class MCFluxWrapper implements ICapabilityProvider, INBTSerializabl
 	}
 
 	@SuppressWarnings("unchecked")
-	public boolean addWrappers(Map<String, ICapabilityProvider> icps) {
+	boolean addWrappers(Map<String, ICapabilityProvider> icps) {
 		int size = icps.size();
 		if (size == 0) {
 			return false;
 		}
 		List<ICapabilityProvider> pl = new ArrayList<>(size);
-		List<String> n = new ArrayList<>(size);
-		List<INBTSerializable<NBTBase>> sl = new ArrayList<>(size);
-		int i = 0;
 		for (Map.Entry<String, ICapabilityProvider> e : icps.entrySet()) {
+			String s = e.getKey().toLowerCase();
 			ICapabilityProvider icp = e.getValue();
-			n.add(i, e.getKey());
-			pl.add(i, icp);
 			if (icp instanceof INBTSerializable)
-				sl.add(i, (INBTSerializable<NBTBase>) icp);
-			else
-				sl.add(i, null);
-			i++;
+				nbts.put(s, (INBTSerializable<NBTBase>) icp);
+			pl.add(icp);
 		}
 		providers = pl.toArray(new ICapabilityProvider[size]);
-		names = n.toArray(new String[size]);
-		serializables = sl.toArray(new INBTSerializable[size]);
+		if (cachedNBT != null) {
+			deserializeNBT(cachedNBT);
+			cachedNBT = null;
+		}
 		return providers.length > 0;
 	}
 
 	@Override public NBTBase serializeNBT() {
 		NBTTagCompound nbttc = new NBTTagCompound();
-		if (names != null && serializables != null)
-			for (int i = 0; i < names.length; i++) {
-				INBTSerializable<NBTBase> nbts = serializables[i];
-				if (nbts == null)
-					continue;
-				nbttc.setTag(names[i], nbts.serializeNBT());
-			}
+		if (nbts.isEmpty())
+			return nbttc;
+		for (Map.Entry<String, INBTSerializable<NBTBase>> e : nbts.entrySet()) {
+			String s = e.getKey();
+			INBTSerializable<NBTBase> nbtz = e.getValue();
+			nbttc.setTag(s, nbtz.serializeNBT());
+		}
 		return nbttc;
 	}
 
 	@Override public void deserializeNBT(NBTBase nbt) {
-		if (nbt != null && nbt instanceof NBTTagCompound && names != null && serializables != null) {
-			NBTTagCompound nbttc = (NBTTagCompound) nbt;
-			for (int i = 0; i < names.length; i++) {
-				INBTSerializable<NBTBase> nbts = serializables[i];
-				if (nbts == null)
-					continue;
-				if (nbttc.hasKey(names[i]))
-					nbts.deserializeNBT(nbttc.getTag(names[i]));
-			}
+		if (providers.length == 0) {
+			cachedNBT = nbt;
+			return;
 		}
+		if (nbt == null || !(nbt instanceof NBTTagCompound))
+			return;
+		NBTTagCompound nbttc = (NBTTagCompound) nbt;
+		for (Map.Entry<String, INBTSerializable<NBTBase>> e : nbts.entrySet())
+			e.getValue().deserializeNBT(nbttc.getTag(e.getKey()));
 	}
 }